requirements.txt
pandas
psycopg2-binary
python-dotenv
pre-commit
boto3
sqlalchemy
black